backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
dj-database-url==2.1.0

# Image Processing
Pillow>=10.0.0
django-cloudinary-storage==0.3.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-decouple==3.8
django-filter==23.5

# Email
sendgrid==6.11.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
